feat: create authentication providers

+++ server/migrations/sqlite/20241112121050_create-provider.sql
@@ -1,6 +1,6 @@
 create table providers (
     application_id text not null
         references applications(id) on delete cascade,
-    kind int not null,
+    kind tinyint not null,
     primary key (application_id, kind)
 );
